+++ app/models/setting.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 # Setting is a configuration element of the platform. Only administrators are allowed to modify Settings
-# For some settings, changing them will involve some callback actions (like rebuilding the stylesheets if the theme color Setting is changed).
+# For some settings, changing them will involve some callback actions (like rebuilding the stylesheets
+# if the theme color Setting has changed).
 # A full history of the previous values is kept in database with the date and the author of the change
 # after_update callback is handled by SettingService
 class Setting < ApplicationRecord
-  has_many :history_values
+  has_many :history_values, dependent: :destroy
   # The following list contains all the settings that can be customized from the Fab-manager's UI.
   # A few of them that are system settings, that should not be updated manually (uuid, origin...).
   validates :name, inclusion:
@@ -151,7 +152,8 @@ class Setting < ApplicationRecord
                              user_change_group
                              user_validation_required
                              user_validation_required_list
-                             show_username_in_admin_list] }
+                             show_username_in_admin_list
+                             store_module] }
   # WARNING: when adding a new key, you may also want to add it in:
   # - config/locales/en.yml#settings
   # - app/frontend/src/javascript/models/setting.ts#SettingName
@@ -188,6 +190,7 @@ class Setting < ApplicationRecord
     previous_value&.created_at
   end
 
+  # @deprecated, prefer Setting.set() instead
   def value=(val)
     admin = User.admins.first
     save && history_values.create(invoicing_profile: admin.invoicing_profile, value: val)
@@ -223,6 +226,6 @@ class Setting < ApplicationRecord
   # Check if the given setting was set
   ##
   def self.set?(name)
-    find_by(name: name)&.value.nil? ? false : true
+    !find_by(name: name)&.value.nil?
   end
 end